agno-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
agno==0.1.0
anthropic==0.7.8
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6 